refactor: CreditCardCompanyModal  파일명 수정
